chimera-brain/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
torch>=2.0.0
transformers>=4.35.0
Pillow>=10.0.0
numpy>=1.24.0
accelerate>=0.25.0
sentencepiece>=0.1.99
redis==5.0.1
redisearch==2.0.0
sentence-transformers==2.2.2